pyproject.toml
[project]
name = "be-aqylshyn-diploma"
version = "0.1.0"
description = ""
authors = [
    {name = "dimash",email = "arystambekdimash005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg-binary (>=3.2.5,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "djangorestframework-jsonapi (>=7.1.0,<8.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-jazzmin (>=3.0.1,<4.0.0)",
    "certifi (>=2025.1.31,<2026.0.0)",
    "django-nested-admin (>=4.1.1,<5.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "openai (>=1.67.0,<2.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "elevenlabs (==1.54.0)",
    "fcm-django (>=2.2.1,<3.0.0)",
    "firebase-admin (>=6.7.0,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
